sbbs_t::gettmplt should be size_t not uint

+++ sbbs3/str.cpp
@@ -2,7 +2,7 @@
 
 /* Synchronet high-level string i/o routines */
 
-/* $Id: str.cpp,v 1.47 2003/07/26 21:40:22 rswindell Exp $ */
+/* $Id: str.cpp,v 1.48 2003/11/02 23:05:36 deuce Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -486,7 +486,7 @@ void sbbs_t::read_sif_dat(char *siffile, char *datfile)
 /* First character MUST be an A,N or !.                                     */
 /* Modes - K_LINE and K_UPPER are supported.                                */
 /****************************************************************************/
-uint sbbs_t::gettmplt(char *strout,char *templt, long mode)
+size_t sbbs_t::gettmplt(char *strout,char *templt, long mode)
 {
 	char	ch,str[256];
 	char	tmplt[128];
